check if file exists before deleting

+++ R/interactive-map/convert.sh
@@ -1,2 +1,6 @@
-rm municipios-centroids.json
+echo "Converting the municipio centroids to GeoJSON"
+if [ -f municipios-centroids.json ]
+then
+    rm municipios-centroids.json
+fi
 ogr2ogr -f "GeoJSON" municipios-centroids.json municipios-centroids.vrt
